feat(0.16Update): Updated mod to function with 0.16 - notes below
Added some very early graphics. It appears that all recipes and tech
information has been lost - all that is left is the basic mod code. Will
need to recreate recipes, tech, ingredient items, and potentially tweak
algorithm.

Algorithm tweak: When an exploding line of trench bombs reaches an
existing body of water, sometimes the final bombs will be skipped over
and not detonate, and also a strange detonation pattern will be added to
the edge of the body of water. Need to look into this. Also,
1-bomb-thick lines of bombs create trenches that can be walked over.
Could potentially leave as-is or make bombs detonate in a 2x2 area.